refactor(storage): move RangeReader and write_range to platform modules
- Move `RangeReader`, `write_range`, `write_range_from_stream`, and
  related helpers out of `content.rs` into platform-specific modules.
- Remove now-unused imports and the shared
  `BUFFER_POOL`/`WritePieceResponse`
  definitions from `content.rs`.
- Strip associated tests from `content.rs` (relocated with the code).
