+++ dragonfly-client-storage/src/content_linux.rs
@@ -233,7 +233,7 @@ impl Content {
         offset: u64,
         length: u64,
         range: Option<Range>,
-    ) -> Result<super::content::RangeReader> {
+    ) -> Result<super::io::RangeReader> {
         let task_path = self.get_task_path(task_id);
 
         // Calculate the target offset and length based on the range.
@@ -246,7 +246,7 @@ impl Content {
             error!("open {:?} failed: {}", task_path, err);
         })?;
 
-        Ok(super::content::RangeReader::new(
+        Ok(super::io::RangeReader::new(
             fd,
             target_offset,
             target_length,
@@ -263,7 +263,7 @@ impl Content {
         offset: u64,
         expected_length: u64,
         stream: &mut S,
-    ) -> Result<super::content::WritePieceResponse>
+    ) -> Result<super::io::WriteRangeResponse>
     where
         S: Stream<Item = std::io::Result<Bytes>> + Unpin + ?Sized,
     {
@@ -278,7 +278,7 @@ impl Content {
                 error!("open {:?} failed: {}", task_path, err);
             })?;
 
-        super::content::write_range_from_stream(
+        super::io::write_range_from_stream(
             fd,
             offset,
             expected_length,
@@ -437,7 +437,7 @@ impl Content {
         offset: u64,
         length: u64,
         range: Option<Range>,
-    ) -> Result<super::content::RangeReader> {
+    ) -> Result<super::io::RangeReader> {
         let task_path = self.get_persistent_task_path(task_id);
 
         // Calculate the target offset and length based on the range.
@@ -450,7 +450,7 @@ impl Content {
             error!("open {:?} failed: {}", task_path, err);
         })?;
 
-        Ok(super::content::RangeReader::new(
+        Ok(super::io::RangeReader::new(
             fd,
             target_offset,
             target_length,
@@ -467,7 +467,7 @@ impl Content {
         offset: u64,
         expected_length: u64,
         reader: &mut R,
-    ) -> Result<super::content::WritePieceResponse> {
+    ) -> Result<super::io::WriteRangeResponse> {
         // Write the piece with positional writes on the cached file descriptor,
         // avoiding reopening and seeking the file for every piece.
         let task_path = self.get_persistent_task_path(task_id);
@@ -479,7 +479,7 @@ impl Content {
                 error!("open {:?} failed: {}", task_path, err);
             })?;
 
-        super::content::write_range(
+        super::io::write_range(
             fd,
             offset,
             expected_length,
@@ -501,7 +501,7 @@ impl Content {
         offset: u64,
         expected_length: u64,
         stream: &mut S,
-    ) -> Result<super::content::WritePieceResponse>
+    ) -> Result<super::io::WriteRangeResponse>
     where
         S: Stream<Item = std::io::Result<Bytes>> + Unpin + ?Sized,
     {
@@ -516,7 +516,7 @@ impl Content {
                 error!("open {:?} failed: {}", task_path, err);
             })?;
 
-        super::content::write_range_from_stream(
+        super::io::write_range_from_stream(
             fd,
             offset,
             expected_length,
@@ -704,7 +704,7 @@ impl Content {
         offset: u64,
         length: u64,
         range: Option<Range>,
-    ) -> Result<super::content::RangeReader> {
+    ) -> Result<super::io::RangeReader> {
         let task_path = self.get_persistent_cache_task_path(task_id);
 
         // Calculate the target offset and length based on the range.
@@ -717,7 +717,7 @@ impl Content {
             error!("open {:?} failed: {}", task_path, err);
         })?;
 
-        Ok(super::content::RangeReader::new(
+        Ok(super::io::RangeReader::new(
             fd,
             target_offset,
             target_length,
@@ -734,7 +734,7 @@ impl Content {
         offset: u64,
         expected_length: u64,
         reader: &mut R,
-    ) -> Result<super::content::WritePieceResponse> {
+    ) -> Result<super::io::WriteRangeResponse> {
         // Write the piece with positional writes on the cached file descriptor,
         // avoiding reopening and seeking the file for every piece.
         let task_path = self.get_persistent_cache_task_path(task_id);
@@ -746,7 +746,7 @@ impl Content {
                 error!("open {:?} failed: {}", task_path, err);
             })?;
 
-        super::content::write_range(
+        super::io::write_range(
             fd,
             offset,
             expected_length,
@@ -769,7 +769,7 @@ impl Content {
         offset: u64,
         expected_length: u64,
         stream: &mut S,
-    ) -> Result<super::content::WritePieceResponse>
+    ) -> Result<super::io::WriteRangeResponse>
     where
         S: Stream<Item = std::io::Result<Bytes>> + Unpin + ?Sized,
     {
@@ -784,7 +784,7 @@ impl Content {
                 error!("open {:?} failed: {}", task_path, err);
             })?;
 
-        super::content::write_range_from_stream(
+        super::io::write_range_from_stream(
             fd,
             offset,
             expected_length,
